printk: ipipe: defer printk() from head domain
The printk() machinery cannot immediately invoke the console driver(s)
when called from the head domain, since such driver code belongs to
the root domain and cannot be shared between domains.

Output issued from the head domain is formatted then logged into a
staging buffer, and a dedicated virtual IRQ is posted to the root
domain for notification. When the virtual IRQ handler runs, the
contents of the staging buffer is flushed to the printk() interface
anew, which may eventually pass the output on to the console drivers
from such a context.